requirements.txt
discord==1.7.3
openai==0.10.1
python-dotenv==0.18.0